requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
redis
celery[redis]
python-dotenv
python-jose[cryptography]
